changed bullet move and player move to floats instead of ints, this was an exstensive edit so had to edit everything from controller to player

+++ core/src/com/mygdx/tanks/Collidable.java
@@ -3,6 +3,6 @@ package com.mygdx.tanks;
 import com.badlogic.gdx.math.Rectangle;
 
 public interface Collidable {
-    public void move(int deltaX, int deltaY);
+    public void move(float deltaX, float deltaY);
     public Rectangle getRectangle();
 }
